mailservice: refactor event handler duplication and tests
